Made cheat 32bit computer compatible

+++ CSGO/CSGO/Source/Features/Miscellaneous/airstuck.cpp
@@ -5,7 +5,7 @@ void CAirstuck::AirStuck( )
 	if ( eng.GetAttack( ) != ( FA_DEFAULT | FA_PRESS ) )
 	{
 		CGlobalVars gvMaxTickCount = eng.GetGlobalVars( );
-		gvMaxTickCount.tickcount = INT_MAX;
+		gvMaxTickCount.tickcount = LONG_MAX;
 		eng.SetGlobalVars( gvMaxTickCount );
 	}
 }
